Added sketch bracket coloring
The code is obviously rough at the moment, but the algorithm choice
and implementation are pretty much completed.

There are several things that I want to improve before merging this
into master:

  - make proper configuration interface

  - add missing docstrings, comments

  - code style cleanup in general

  - term cleanup in general (e.g., extents? maybe they are 'spans')

  - implement merging of the consecutive lines

After this, I believe, I will leave the actual coloring as a stub
and move on to configuration jobs: reading, parsing, processing,
generating color scopes, etc.