Added basic functions for updating scope colors in theme files
This is the only way for a Sublime plugin to do all coloring by itself:
manually update the relevant configuration files. Thus we can compute and
write down any colors we want. Not a clean solution, but meh... there is
no legal API for this.

Additional scopes are just added and left there as I have no idea when
they are to be removed. They are (presumably) unique for LispHighlighter,
so this should not be a problem if some random unused scopes are there in
the theme file.

Also need function to parse the theme file and extract colors for the
background and the current line. And probably some settings listener to
update the colors whenever the current theme is changed by the user.

I am not using a proper XML parser for this. It is easier to just assume
sane config file format that do _everything_ right. Maybe some day this
won't be true and I will have to add one.