Added example configuration file
Finally, the feature list is complete, yay! \o/

There are some things that are missing at the moment. The most prominent
one is that scope type overriding does not work as described. Adjacency
and inconsistency are types of a scope, not traits; i.e., a scope cannot
be primary mainline and inconsistent. It will be 'just inconsistent'.
However, if inconsistent highlighting is disabled, it _should_ be treated
as a primary one for the highlighting needs.

Obviosly, all override machinery is not implemented also. These things
will be done in separate branches. Here I want to concentrate on config
parsing and validation. Also I want to make a handy Configuration class
that transparently maps all the settings to Python objects. __metashit__
will surely ensue.