Added main configuration file pattern
I guess it will be easier to describe the configuration in a declarative
way and implement a type-checking default-replacing recursive unifier,
rather than try to do the same in an imperative manner with tons of
explicit checks. Not to mention that this will be *much more* readable
and maintainable.

Need to think over how to squeeze some additional checks in there;
e.g., to verify that only "color" or "colors" is used in scope settings.